apres p

+++ libft/ft_power.c
@@ -6,16 +6,16 @@
 /*   By: amineau <amineau@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/02/26 02:57:38 by amineau           #+#    #+#             */
-/*   Updated: 2016/02/26 02:59:38 by amineau          ###   ########.fr       */
+/*   Updated: 2016/02/26 08:06:35 by amineau          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_power(int nbr, int base)
+uintmax_t	ft_power(int nbr, int base)
 {
-	int i;
-	int res;
+	int			i;
+	uintmax_t	res;
 
 	i = 0;
 	res = 1;
